chore: rename proto to peer
